refactor : 게임결과 조건식 이름 부여
