--- src/main/java/blackjack/card/domain/result/LoseStrategy.java
@@ -1,18 +1,0 @@
-package blackjack.card.domain.result;
-
-import blackjack.card.domain.CardBundle;
-
-public class LoseStrategy extends GameResultStrategy {
-
-    private static final int LOSE_VALUE = -1;
-
-    @Override
-    protected boolean enough(CardBundle gamblerCardBundle, CardBundle dealerCardBundle) {
-        return gamblerCardBundle.isBurst() && !dealerCardBundle.isBurst() || dealerCardBundle.isBlackjack() && !gamblerCardBundle.isBlackjack();
-    }
-
-    @Override
-    protected boolean isMatch(int compare) {
-        return LOSE_VALUE == compare;
-    }
-}
